hotfix: WebConfig 패키지 이동 및 CORS 허용 도메인 추가
- CORS 허용 도메인에 https://api.pickeat.io.kr 추가하여 API 접근 가능 도메인 확장
